problem solved by increase the stack limit

+++ rt_next_week/cuda/rtweekend.h
@@ -14,8 +14,8 @@ using std::sqrt;
 
 // Constants
 
-const float infinity = std::numeric_limits<double>::infinity();
-const float pi = 3.1415926535897932385;
+const float infinity = std::numeric_limits<float>::infinity();
+const float pi = 3.1415926535897932385f;
 
 // Utility Functions
